# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API Softphone",
        "description": "",
        "contact": {
            "name": "Daniel Lopes Manfrini",
            "email": "analistadev.04@plansul.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/uras/get": {
            "get": {
                "tags": [
                    "Uras"
                ],
                "summary": "Coleta as uras.",
                "operationId": "cbded271ca0482315fbd87334a08c7f4",
                "parameters": [
                    {
                        "name": "exten",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UrasGetResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/uras/create": {
            "post": {
                "tags": [
                    "Uras"
                ],
                "summary": "Realiza o cadastro da ura.",
                "operationId": "cbea09ca29e475fcc20e9521aaa6b042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrasCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/uras/edit": {
            "put": {
                "tags": [
                    "Uras"
                ],
                "summary": "Realiza a edição da ura.",
                "operationId": "b47096afc8c09134c47faa724496fc8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UrasEditRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/uras/delete": {
            "delete": {
                "tags": [
                    "Uras"
                ],
                "summary": "Realiza a exclusao da ura.",
                "operationId": "e3b98ca037511a443d193bfc68eff034",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UrasGetResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "context": {
                        "type": "string",
                        "example": "ura"
                    },
                    "exten": {
                        "type": "any",
                        "example": "3000,switch-2"
                    },
                    "priority": {
                        "type": "integer",
                        "example": 5
                    },
                    "app": {
                        "type": "string",
                        "example": 1
                    },
                    "appdata": {
                        "type": "string",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UrasCreateRequest": {
                "required": [
                    "context",
                    "exten",
                    "priority",
                    "app",
                    "appdata"
                ],
                "properties": {
                    "context": {
                        "type": "string",
                        "example": "ura"
                    },
                    "exten": {
                        "type": "any",
                        "example": "3000,switch-2"
                    },
                    "priority": {
                        "type": "integer",
                        "example": 5
                    },
                    "app": {
                        "type": "string",
                        "example": 1
                    },
                    "appdata": {
                        "type": "string",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "UrasEditRequest": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "context": {
                        "type": "string",
                        "example": "ura"
                    },
                    "exten": {
                        "type": "any",
                        "example": "3000,switch-2"
                    },
                    "priority": {
                        "type": "integer",
                        "example": 5
                    },
                    "app": {
                        "type": "string",
                        "example": 1
                    },
                    "appdata": {
                        "type": "string",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    }
}